requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
captcha==0.7.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
distro==1.9.0
extra-streamlit-components==0.1.80
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
groq==0.24.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.37.1
numpy==2.2.5
openai==1.77.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.45.0
streamlit-authenticator==0.4.2
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
